+++ src/_locales/en/options.ts
@@ -381,7 +381,9 @@ export const locale: typeof _locale = {
       trans_tw: 'for traditional Chinese translation',
       weblio: 'for helping add Weblio dict',
       machine_translators:
-        'for implementing Alibaba Translate, Niutrans and Volcengine Translate'
+        'for implementing Alibaba Translate, Niutrans and Volcengine Translate',
+      bingtrans: 'for implementing Bing Translate',
+      trans_ko: 'for the Korean translation'
     },
     contact_author: 'Contact Author',
     donate: 'Donate',