+++ src/_locales/ko/background.ts
@@ -1,0 +1,11 @@
+import { locale as _locale } from '../zh-CN/background'
+
+export const locale: typeof _locale = {
+  app: {
+    off: 'Saladict가 비활성화되었습니다. (독립 실행 패널은 계속 사용 가능)',
+    tempOff:
+      '현재 탭에서 Saladict가 비활성화되었습니다. (독립 실행 패널은 계속 사용 가능)',
+    unsupported:
+      '현재 탭에서는 내장형 Saladict 패널을 사용할 수 없습니다. 독립 실행 패널을 이용해 주세요.'
+  }
+}