+++ src/components/dictionaries/bingtrans/_locales.ts
@@ -1,0 +1,8 @@
+import { getMachineLocales } from '../locales'
+
+export const locales = getMachineLocales({
+  en: 'Bing Translate',
+  'zh-CN': '必应翻译',
+  'zh-TW': '必應翻譯',
+  ko: 'Bing 번역'
+})