+++ src/components/dictionaries/deeplx/_locales.ts
@@ -3,5 +3,6 @@ import { getMachineLocales } from '../locales'
 export const locales = getMachineLocales({
   en: 'DeepLX',
   'zh-CN': 'DeepLX',
-  'zh-TW': 'DeepLX'
+  'zh-TW': 'DeepLX',
+  ko: 'DeepLX'
 })