+++ src/components/dictionaries/youdaotrans/_locales.ts
@@ -3,5 +3,6 @@ import { getMachineLocales } from '../locales'
 export const locales = getMachineLocales({
   en: 'Youdao Translate',
   'zh-CN': '有道翻译',
-  'zh-TW': '有道翻譯'
+  'zh-TW': '有道翻譯',
+  ko: '유다오 번역'
 })